# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Expresso API",
    "description": "Expresso API"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "name": "localhost"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Welcome to Expresso API"],
        "summary": "Welcome Endpoint",
        "description": "Welcome to Expresso API.",
        "responses": {
          "200": {
            "description": "Welcome Endpoint"
          }
        }
      }
    },
    "/api/employees": {
      "get": {
        "tags": ["Employees"],
        "summary": "Get all employees",
        "description": "Get all employees",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "tags": ["Employees"],
        "summary": "Create a new employee",
        "description": "Create a new employee",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Employee object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/employees/{id}": {
      "get": {
        "tags": ["Employees"],
        "summary": "Get employee by id",
        "description": "Get employee by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Employee id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "put": {
        "tags": ["Employees"],
        "summary": "Update employee by id",
        "description": "Update employee by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Employee id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Employee object that needs to be updated in the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["Employees"],
        "summary": "Delete employee by id",
        "description": "Delete employee by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Employee id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/employees/{id}/timesheets": {
      "get": {
        "tags": ["Timesheets"],
        "summary": "Get employee timesheet by id",
        "description": "Get employee timesheet by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Employee id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "tags": ["Timesheets"],
        "summary": "Create a new timesheet",
        "description": "Create a new timesheet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Employee id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Timesheet object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Timesheet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/employees/{employeeId}/timesheets/{timesheetId}": {
      "get": {
        "tags": ["Timesheets"],
        "summary": "Get timesheet by id",
        "description": "Get timesheet by id",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "description": "Employee id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "timesheetId",
            "in": "path",
            "description": "Timesheet id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "put": {
        "tags": ["Timesheets"],
        "summary": "Update timesheet by id",
        "description": "Update timesheet by id",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "description": "Employee id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "timesheetId",
            "in": "path",
            "description": "Timesheet id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Timesheet object that needs to be updated in the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Timesheet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["Timesheets"],
        "summary": "Delete timesheet by id",
        "description": "Delete timesheet by id",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "description": "Employee id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "timesheetId",
            "in": "path",
            "description": "Timesheet id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/menus": {
      "get": {
        "tags": ["Menu"],
        "summary": "Get all menus",
        "description": "Get all menus",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "tags": ["Menu"],
        "summary": "Create a new menu",
        "description": "Create a new menu",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Menu object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/menus/{id}": {
      "get": {
        "tags": ["Menu"],
        "summary": "Get menu by id",
        "description": "Get menu by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Menu id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "put": {
        "tags": ["Menu"],
        "summary": "Update menu by id",
        "description": "Update menu by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Menu id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Menu object that needs to be updated in the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["Menu"],
        "summary": "Delete menu by id",
        "description": "Delete menu by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Menu id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/menus/{menuId}/menu-items": {
      "get": {
        "tags": ["Menu Items"],
        "summary": "Get all menu items",
        "description": "Get all menu items",
        "parameters": [
          {
            "name": "menuId",
            "in": "path",
            "description": "Menu id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "tags": ["Menu Items"],
        "summary": "Create a new menu item",
        "description": "Create a new menu item",
        "parameters": [
          {
            "name": "menuId",
            "in": "path",
            "description": "Menu id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Menu item object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MenuItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/menus/{menuId}/menu-items/{menuItemId}": {
      "get": {
        "tags": ["Menu Items"],
        "summary": "Get menu item by id",
        "description": "Get menu item by id",
        "parameters": [
          {
            "name": "menuId",
            "in": "path",
            "description": "Menu id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "menuItemId",
            "in": "path",
            "description": "Menu item id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "put": {
        "tags": ["Menu Items"],
        "summary": "Update menu item by id",
        "description": "Update menu item by id",
        "parameters": [
          {
            "name": "menuId",
            "in": "path",
            "description": "Menu id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "menuItemId",
            "in": "path",
            "description": "Menu item id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Menu item object that needs to be updated in the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MenuItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["Menu Items"],
        "summary": "Delete menu item by id",
        "description": "Delete menu item by id",
        "parameters": [
          {
            "name": "menuId",
            "in": "path",
            "description": "Menu id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "menuItemId",
            "in": "path",
            "description": "Menu item id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    }
  },
  "definitions": {
    "Employee": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "position": {
          "type": "string"
        },
        "wage": {
          "type": "string"
        },
        "is_current_employee": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "Timesheet": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date"
        },
        "hours": {
          "type": "integer",
          "format": "int32"
        },
        "rate": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Menu": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        }
      }
    },
    "MenuItem": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "inventory": {
          "type": "integer"
        },
        "price": {
          "type": "integer"
        }
      }
    }
  }
}